check process exist before kill

+++ entrypoint.sh
@@ -11,7 +11,9 @@ while true; do
     fn="/opt/nmon/$(date +%y%m%d_%H%M).nmon"
     pd="$(/opt/nmon/nmon -F $fn -s $s -c $c -T -p)"
     { echo "nmon p:default e:`hostname` f:`hostname`_file.nmon"; tail -f $fn --pid=$pd; } | nc $atsdUrl $atsdPort
-    kill -9 $pd
+    if kill -0 $pd 2>/dev/null; then
+        kill -9 $pd
+    fi
     sleep 10
 done
